==> 01_create_sample.ipynb <==
import pandas as pd

number_of_items = 150

retail_goods_and_prices_csv = pd.read_csv('./data/RetailGoodsAndPrices.csv', sep='\t')
retail_goods_and_prices_sample = retail_goods_and_prices_csv.sample(number_of_items)
retail_goods_and_prices_sample = retail_goods_and_prices_sample.drop('Code', axis=1)
retail_goods_and_prices_sample.reset_index(drop=True, inplace=True)
retail_goods_and_prices_sample['Code'] = retail_goods_and_prices_sample.index + 1
retail_goods_and_prices_sample.to_csv('./data/RetailGoodsAndPricesSample.csv', index=False, sep='\t', columns=['Code', 'Group', 'Subgroup', 'Brand', 'Item', 'Description', 'Wholesale price (USD)'])
