==> 03_set_retail_prices.ipynb <==
import pandas as pd

markup_percentage = 5 # %

retail_goods_and_prices_csv = pd.read_csv('./normalized/05_wholesale_prices.csv', sep='\t')
retail_goods_and_prices_csv['Retail price (USD) Raw'] = retail_goods_and_prices_csv['Wholesale price (USD)'] * (1 + 0.01 * markup_percentage)
retail_goods_and_prices_csv['Retail price (USD)'] = retail_goods_and_prices_csv['Retail price (USD) Raw'].round(2)

retail_goods_and_prices_csv.to_csv('./normalized/06_retail_prices.csv', index=False, sep='\t', columns=['Code', 'Retail price (USD)'])
