==> 05_discount_cards.ipynb <==
import pandas as pd
import random

orders = pd.read_csv('./orders/orders.csv', sep='\t')
discount_cards = pd.read_csv('./data/Discount_cards.csv')

orders_with_discount_cards_columns = ['OrderID', 'DiscountCardID']
orders_with_discount_cards = pd.DataFrame(columns=orders_with_discount_cards_columns)

for index, row in orders.iterrows():
    order_id = row['ID']
    
    discount_card_id = discount_cards['id'][random.randint(0, len(discount_cards) - 1)] if random.randint(1, random.randint(1, 25)) < 4 else 0
    
    if discount_card_id != 0:
        new_order_discount_card_record = pd.DataFrame([{
            'OrderID': order_id, 
            'DiscountCardID': discount_card_id}])
        orders_with_discount_cards = pd.concat([orders_with_discount_cards, new_order_discount_card_record], ignore_index=True)

orders_with_discount_cards.to_csv('./orders/discount_cards.csv', sep='\t', index=False, columns=orders_with_discount_cards_columns)
