==> 06_join_all_data.ipynb <==
import pandas as pd
import numpy as np

orders = pd.read_csv('./orders/orders.csv', sep='\t')
subgroups = pd.read_csv('./normalized/02_subgroups.csv', sep='\t')
items = pd.read_csv('./normalized/04_items.csv', sep='\t')
retail_prices = pd.read_csv('./normalized/06_retail_prices.csv', sep='\t')
discount_cards = pd.read_csv('./orders/discount_cards.csv', sep='\t')
discount_cards['DiscountCardID'] = pd.to_numeric(discount_cards['DiscountCardID'])
discount_cards = discount_cards.astype(np.int64)

joined = pd.merge(orders, items, how='left', left_on=['ItemCode'], right_on=['Code'])
joined = pd.merge(joined, subgroups, how='inner', left_on=['Subgroup_ID'], right_on=['ID'])
joined = pd.merge(joined, retail_prices, how='inner', left_on=['ItemCode'], right_on=['Code'])
joined = pd.merge(joined, discount_cards, how='left', left_on=['ID_x'], right_on=['OrderID'])
joined.drop(['ID_y', 'Code_x', 'Item', 'Description', 'Subgroup', 'Code_y', 'OrderID'], axis=1, inplace=True) 
joined['DiscountCardID'] = pd.to_numeric(joined['DiscountCardID'])
joined.rename(columns={'ID_x': 'ID'},  inplace=True)
joined.reset_index(drop=True, inplace=True)
joined.sort_values(by=['ID'], ascending=True, inplace=True)
joined.to_csv('./orders/orders_full.csv', index=False, sep='\t')
